requirements.txt
hikari
pyyaml
sqlalchemy